Daily Transactions Dataset
About Dataset

The "Daily Transactions" dataset contains information on dummy transactions made by an individual on a daily basis. The dataset includes data on the products that were purchased, the amount spent on each product, the date and time of each transaction, the payment mode of each transaction, and the source of each record (Expense/Income).

This dataset can be used to analyze purchasing behavior and money management, forecasting expenses, and optimizing savings and budgeting strategies. The dataset is well-suited for data analysis and machine learning applications,it can be used to train predictive models and make data-driven decisions.

Column Descriptors

    Date: The date and time when the transaction was made
    Mode: The payment mode used for the transaction
    Category: Each record is divided into a set of categories of transactions
    Subcategory: Categories are further broken down into Subcategories of transactions
    Note: A brief description of the transaction made
    Amount: The transactional amount
    Income/Expense: The indicator of each transaction representing either expense or income
    Currency: All transactions are recorded in official currency of India